Django/noticias_upsa/requirements.txt
django==4.0
ipython==8.1.1
djangorestframework==3.13.1
